Handle bug in OSX grep
There is a bug in OSX grep when it's invoked with `-n` flag.

Notice the last line in the below example output does *not* have a
number. That a bug.

  168: a/copycat.tmux
  170: a/copycat.tmux
   b/copycat.tmux

To compensate for that - if a line does not have a number (bug case),
we're fetching a number from a previous line.