Fix OSX awk bug where `=` is first variable char
OS X version of awk program does not permit `=` to be the first
character in the variable. The following example fails on OS X 10.9.3:

    awk -v a="=test" -v b="est" 'BEGIN{print index(a,b)}'

(notice the variable a content starts with `=`).
To prevent this and enable somewhat correct results, `=` character, when
at the beginning of the line, is changed to `.`

Fixes #24