Test suite runs vi *and* emacs copy-mode too
There are differences between vi and emacs copy modes in tmux. We're
even handling them differently in tmux-copycat.

Because of this, there's a need to run the tests in both the vi and
emacs tmux copy mode.
This is achieved by setting the EDITOR variable in test runner script.

Tmux picks up the EDITOR var and automatically sets copy-mode based on
it's value.

Closes #38