make working directory only be accessible by user
This makes /tmp/tmux-NNNN-copycat/ only accessible by the current user
of the tmux session. While not important on single user systems, this
could be a concern on multi-user servers, where using tmux-copycat would
capture the scrollback text of the current terminal to a file that could
be read by other users, especially if you don't have a strict umask.