BUG: keybinding values undergo globbing, may break mappings
I have configured the copycat search key to '?' (set -g @copycat_search '?' in .tmux.conf), and all of a sudden the mapping was assigned to prefix + 6 instead of prefix + ?. Turned out I had a file ~/6, and the ? mistakenly is expanded to 6.
The problem is the unquoted variable reference "for key in $key_bindings". As we need the word splitting for the (undocumented?) feature of allowing multiple key mappings, it cannot be quoted.
As a fix, use read with a here-string to perform word splitting in an array variable, and then iterate over the array elements.