Improve performance by binding copycat keys just once
All of the copycat scripts check to make sure we're in copycat mode, and
if we're not they are essentially a no-op. This means we can safely
leave the keys bound to those scripts even when we're not in copycat
mode.

The benefit of this is that we don't have to bind and unbind the keys
every time we enter or leave copycat mode, which helps improve
performance.